requirements.txt
altair==5.5.0
anyio==4.12.1
apify_client==2.5.0
apify_shared==2.2.0
APScheduler==3.11.2
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudscraper==1.2.71
colorama==0.4.6
cryptography==46.0.5
cssselect==1.4.0
curl_cffi==0.14.0
DataRecorder==3.6.2
DownloadKit==2.0.7
DrissionPage==4.1.1.2
et_xmlfile==2.0.0
feedparser==6.0.12
filelock==3.25.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-auth-oauthlib==1.3.0
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
impit==0.12.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
narwhals==2.18.0
numpy==1.26.4
oauthlib==3.3.1
openpyxl==3.1.5
packaging==23.2
pandas==2.2.1
pillow==10.4.0
protobuf==4.25.8
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.30.0
rsa==4.9.1
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
streamlit==1.32.0
tenacity==8.5.0
tldextract==5.3.1
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
websocket-client==1.9.0
